test: TECHDEBT-33 — split test_handlers.py into per-module use-case test files
Move all 44 remaining test methods from the monolithic
test/api/v2/backend/test_handlers.py into per-module files under
test/core/use_cases/, mirroring the source layout:

- test_embargo_use_cases.py  (TestEmbargoUseCases — 11 tests)
- test_note_use_cases.py     (TestNoteUseCases — 6 tests)
- test_status_use_cases.py   (TestStatusUseCases — 7 tests)
- test_case_use_cases.py     (TestCaseUseCases — 6 tests)
- test_report_use_cases.py   (TestReportReceiptRM — 3 tests added)

Delete test/api/v2/backend/test_handlers.py.

976 tests pass (was 913; +63 reflects newly visible migrated tests).
